Continued the models refactor by putting the baseline classes into their own file; Put SklearnBase into base_models.py
