+++ dna/models/models.py
@@ -2,20 +2,24 @@ import json
 import os
 import typing
 
-import autosklearn.regression as autosklearn
 import numpy as np
 import pandas as pd
-from sklearn import linear_model
 import torch
 
 from dna import utils
 from dna.data import Dataset, GroupDataLoader, PMFDataset, RNNDataLoader, group_json_objects
-from dna.kND import KNearestDatasets
 from .base_models import PyTorchModelBase
 from .base_models import PyTorchRegressionRankSubsetModelBase
 from .base_models import RegressionModelBase
 from .base_models import RankModelBase
 from .base_models import SubsetModelBase
+from .baselines import MedianBaseline
+from .baselines import MeanBaseline
+from .baselines import PerPrimitiveBaseline
+from .baselines import AutoSklearnMetalearner
+from .baselines import LinearRegressionBaseline
+from .baselines import RandomBaseline
+from .baselines import MetaAutoSklearn
 from .torch_modules.dna_module import DNAModule
 from .torch_modules.dag_lstm_mlp import DAGLSTMMLP
 from .torch_modules.hidden_mlp_dag_lstm_mlp import HiddenMLPDAGLSTMMLP
@@ -234,325 +238,6 @@ class HiddenDAGLSTMRegressionModel(DAGLSTMRegressionModel):
             device=self.device,
             seed=self._model_seed,
         )
-
-
-class MeanBaseline(RegressionModelBase):
-
-    def __init__(self, seed=0):
-        super().__init__(seed=seed)
-        self.mean = None
-
-    def fit(self, data, *, validation_data=None, output_dir=None, verbose=False):
-        total = 0
-        for instance in data:
-            total += instance['test_f1_macro']
-        self.mean = total / len(data)
-        self.fitted = True
-
-    def predict_regression(self, data, *, verbose=False):
-        if self.mean is None:
-            raise ModelNotFitError('MeanBaseline not fit')
-        return [self.mean] * len(data)
-
-
-class MedianBaseline(RegressionModelBase):
-
-    def __init__(self, seed=0):
-        super().__init__(seed=seed)
-        self.median = None
-
-    def fit(self, data, *, validation_data=None, output_dir=None, verbose=False):
-        self.median = np.median([instance['test_f1_macro'] for instance in data])
-        self.fitted = True
-
-    def predict_regression(self, data, *, verbose=False):
-        if self.median is None:
-            raise ModelNotFitError('MeanBaseline not fit')
-        return [self.median] * len(data)
-
-
-class PerPrimitiveBaseline(RegressionModelBase, RankModelBase, SubsetModelBase):
-
-    def __init__(self, seed=0):
-        super().__init__(seed=seed)
-        self.primitive_scores = None
-
-    def fit(self, data, *, validation_data=None, output_dir=None, verbose=False):
-        # for each primitive, get the scores of all the pipelines that use the primitive
-        primitive_score_totals = {}
-        for instance in data:
-            for primitive in instance['pipeline']['steps']:
-                if primitive['name'] not in primitive_score_totals:
-                    primitive_score_totals[primitive['name']] = {
-                        'total': 0,
-                        'count': 0,
-                    }
-                primitive_score_totals[primitive['name']]['total'] += instance['test_f1_macro']
-                primitive_score_totals[primitive['name']]['count'] += 1
-
-        # compute the average pipeline score per primitive
-        self.primitive_scores = {}
-        for primitive_name in primitive_score_totals:
-            total = primitive_score_totals[primitive_name]['total']
-            count = primitive_score_totals[primitive_name]['count']
-            self.primitive_scores[primitive_name] = total / count
-
-        self.fitted = True
-
-    def predict_regression(self, data, **kwargs):
-        if self.primitive_scores is None:
-            raise ModelNotFitError('PerPrimitiveBaseline not fit')
-
-        predictions = []
-        for instance in data:
-            prediction = 0
-            for primitive in instance['pipeline']['steps']:
-                prediction += self.primitive_scores[primitive['name']]
-            prediction /= len(instance['pipeline']['steps'])
-            predictions.append(prediction)
-
-        return predictions
-
-    def predict_rank(self, data, **kwargs):
-        predictions = self.predict_regression(data, **kwargs)
-        ranks = list(utils.rank(predictions))
-        return {
-            'pipeline_id': [instance['pipeline_id'] for instance in data],
-            'rank': ranks,
-        }
-
-    def predict_subset(self, data, k, **kwargs):
-        if not self.fitted:
-            raise Exception('model not fit')
-
-        ranked_data = self.predict_rank(data, **kwargs)
-        top_k = pd.DataFrame(ranked_data).nsmallest(k, columns='rank')['pipeline_id']
-        return top_k.tolist()
-
-
-class RandomBaseline(RankModelBase):
-
-    def __init__(self, seed=0):
-        super().__init__(seed=seed)
-        self._random_state = np.random.RandomState(seed)
-        self.fitted = True
-
-    def fit(self, *args, **kwargs):
-        pass
-
-    def predict_rank(self, data, *, verbose=False):
-        predictions = list(range(len(data)))
-        self._random_state.shuffle(predictions)
-        return {
-            'pipeline_id': [instance['pipeline_id'] for instance in data],
-            'rank': predictions,
-        }
-
-    def predict_subset(self, data, k, **kwargs):
-        predictions = self._random_state.choice(data, k)
-        return [instance['pipeline_id'] for instance in predictions]
-
-
-class SklearnBase(RegressionModelBase, RankModelBase, SubsetModelBase):
-
-    def __init__(self, seed=0):
-        super().__init__(seed=seed)
-        self.pipeline_key = 'pipeline'
-        self.steps_key = 'steps'
-        self.prim_name_key = 'name'
-
-    def fit(self, data, *, validation_data=None, output_dir=None, verbose=False):
-        self.one_hot_primitives_map = self._one_hot_encode_mapping(data)
-        data = pd.DataFrame(data)
-        y = data['test_f1_macro']
-        X_data = self.prepare_data(data)
-        self.regressor.fit(X_data, y)
-        self.fitted = True
-
-    def predict_regression(self, data, *, verbose=False):
-        if not self.fitted:
-            raise ModelNotFitError('{} not fit'.format(type(self).__name__))
-
-        data = pd.DataFrame(data)
-        X_data = self.prepare_data(data)
-        return self.regressor.predict(X_data).tolist()
-
-    def predict_rank(self, data, *, verbose=False):
-        if not self.fitted:
-            raise ModelNotFitError('{} not fit'.format(type(self).__name__))
-
-        predictions = self.predict_regression(data)
-        ranks = utils.rank(predictions)
-        return {
-            'pipeline_id': [instance['pipeline_id'] for instance in data],
-            'rank': ranks,
-        }
-
-    def predict_subset(self, data, k, **kwargs):
-        if not self.fitted:
-            raise Exception('model not fit')
-
-        ranked_data = self.predict_rank(data, **kwargs)
-        top_k = pd.DataFrame(ranked_data).nsmallest(k, columns='rank')['pipeline_id']
-        return top_k.tolist()
-
-    def prepare_data(self, data):
-        # expand the column of lists of metafeatures into a full dataframe
-        metafeature_df = pd.DataFrame(data.metafeatures.values.tolist()).reset_index(drop=True)
-        assert np.isnan(metafeature_df.values).sum() == 0, 'metafeatures should not contain nans'
-        assert np.isinf(metafeature_df.values).sum() == 0, 'metafeatures should not contain infs'
-
-        encoded_pipelines = self.one_hot_encode_pipelines(data)
-        assert np.isnan(encoded_pipelines.values).sum() == 0, 'pipeline encodings should not contain nans'
-        assert np.isinf(encoded_pipelines.values).sum() == 0, 'pipeline encodings should not contain infs'
-
-        # concatenate the parts together and validate
-        assert metafeature_df.shape[0] == encoded_pipelines.shape[0], 'number of metafeature instances does not match number of pipeline instances'
-        X_data = pd.concat([encoded_pipelines, metafeature_df], axis=1, ignore_index=True)
-        assert X_data.shape[1] == (encoded_pipelines.shape[1] + metafeature_df.shape[1]), 'dataframe was combined incorrectly'
-        return X_data
-
-    def _one_hot_encode_mapping(self, data):
-        primitive_names = set()
-
-        # Get a set of all the primitives in the train set
-        for instance in data:
-            primitives = instance[self.pipeline_key][self.steps_key]
-            for primitive in primitives:
-                primitive_name = primitive[self.prim_name_key]
-                primitive_names.add(primitive_name)
-
-        primitive_names = sorted(primitive_names)
-
-        # Get one hot encodings of all the primitives
-        self.n_primitives = len(primitive_names)
-        encoding = np.identity(n=self.n_primitives)
-
-        # Create a mapping of primitive names to one hot encodings
-        primitive_name_to_enc = {}
-        for (primitive_name, primitive_encoding) in zip(primitive_names, encoding):
-            primitive_name_to_enc[primitive_name] = primitive_encoding
-
-        return primitive_name_to_enc
-
-    def one_hot_encode_pipelines(self, data):
-        return pd.DataFrame([self.encode_pipeline(pipeline) for pipeline in data[self.pipeline_key]])
-
-    def encode_pipeline(self, pipeline):
-        """
-        Encodes a pipeline by OR-ing the one-hot encoding of the primitives.
-        """
-        encoding = np.zeros(self.n_primitives)
-        for primitive in pipeline[self.steps_key]:
-            primitive_name = primitive[self.prim_name_key]
-            # get the position of the one hot encoding
-            primitive_index = np.argmax(self.one_hot_primitives_map[primitive_name])
-            encoding[primitive_index] = 1
-        return encoding
-
-
-class LinearRegressionBaseline(SklearnBase):
-
-    def __init__(self, seed=0):
-        super().__init__(seed=seed)
-        self.regressor = linear_model.LinearRegression()
-        self.fitted = False
-
-
-class MetaAutoSklearn(SklearnBase):
-
-    def __init__(self, time_left_for_this_task=60, per_run_time_limit=10, seed=0):
-        super().__init__(seed=seed)
-        self.regressor = autosklearn.AutoSklearnRegressor(
-            time_left_for_this_task=time_left_for_this_task, per_run_time_limit=per_run_time_limit, seed=seed
-        )
-        self.fitted = False
-
-
-class AutoSklearnMetalearner(RegressionModelBase, RankModelBase, SubsetModelBase):
-
-    def __init__(self, seed=0):
-        super().__init__(seed=seed)
-        self._knd = KNearestDatasets(metric='l1')
-
-    def _predict(self, data, method, k=None):
-        data = pd.DataFrame(data)
-        # they all should have the same dataset and metafeatures so take it from the first row
-        dataset_metafeatures = pd.Series(data['metafeatures'].iloc[0])
-        queried_pipelines = data['pipeline_id']
-
-        if method == 'all':
-            predicted_pipelines = self._knd.knn_regression(dataset_metafeatures)
-            predicted_pipelines = predicted_pipelines.sort_values(ascending=False).index.tolist()
-        elif method == 'k':
-            predicted_pipelines = self._knd.kBestSuggestions(dataset_metafeatures, k=k)
-        else:
-            raise ValueError('Unknown method: {}'.format(method))
-
-        for pipeline_id in set(predicted_pipelines).difference(set(queried_pipelines)):
-            predicted_pipelines.remove(pipeline_id)
-
-        return predicted_pipelines
-
-    def predict_regression(self, data, **kwargs):
-        predictions = []
-        cached_predictions = {}
-        for instance in data:
-            dataset_id = instance['dataset_id']
-            if dataset_id not in cached_predictions:
-                metafeatures = pd.Series(instance['metafeatures'])
-                cached_predictions[dataset_id] = self._knd.knn_regression(metafeatures)
-            pipeline_id = instance['pipeline_id']
-            predictions.append(cached_predictions[dataset_id].get(pipeline_id, None))
-        return predictions
-
-    def predict_subset(self, data, k, **kwargs):
-        """
-        Recommends at most k pipelines from data expected to perform well for the provided dataset.
-        """
-        return self._predict(data, method='k', k=k)
-
-    def predict_rank(self, data, **kwargs):
-        """
-        Ranks all pipelines in data.
-        """
-        ranked_pipelines = self._predict(data, method='all')
-
-        assert len(ranked_pipelines) == len(data), '{} {}'.format(len(ranked_pipelines), len(data))
-
-        return {
-            'pipeline_id': ranked_pipelines,
-            'rank': list(range(len(ranked_pipelines))),
-        }
-
-    def fit(self, train_data, *args, **kwargs):
-        self._runs = self._process_runs(train_data)
-        self._metafeatures = self._process_metafeatures(train_data)
-        self._knd.fit(self._metafeatures, self._runs)
-        self.fitted = True
-
-    @staticmethod
-    def _process_runs(data):
-        """
-        This function is used to transform the dataframe into a workable object fot the KNN, with rows of pipeline_ids
-        and columns of datasets, with the inside being filled with the scores
-        :return:
-        """
-        new_runs = {}
-        for index, row in enumerate(data):
-            dataset_name = row['dataset_id']
-            if dataset_name not in new_runs:
-                new_runs[dataset_name] = {}
-            new_runs[dataset_name][row['pipeline_id']] = row['test_f1_macro']
-        final_new = pd.DataFrame(new_runs)
-        return final_new
-
-    @staticmethod
-    def _process_metafeatures(data):
-        metafeatures = pd.DataFrame(data)[['dataset_id', 'metafeatures']].set_index('dataset_id')
-        metafeatures = pd.DataFrame(metafeatures['metafeatures'].tolist(), metafeatures.index)
-        metafeatures.drop_duplicates(inplace=True)
-        return metafeatures
 
 
 class ProbabilisticMatrixFactorization(PyTorchRegressionRankSubsetModelBase):
